# Conflicts:
#	src/api/auth.js
#	src/api/index.js
#	src/api/posts.js
#	src/components/posts/PostListItem.vue

+++ src/components/posts/PostListItem.vue
@@ -8,7 +8,7 @@
     </div>
     <div class="post-time">
       {{ postItem.createdAt }}
-      <i class="icon ion-md-create"></i>
+      <i class="icon ion-md-create" @click="routeEditPage"></i>
       <i class="icon ion-md-trash" @click="deleteItem"></i>
     </div>
   </li>
@@ -16,6 +16,7 @@
 
 <script>
 import { deletePost } from '@/api/posts';
+
 export default {
   props: {
     postItem: {
@@ -25,9 +26,15 @@ export default {
   },
   methods: {
     async deleteItem() {
-      // this.postItem._id 으로 부모 모듈로 부터 받은 값에 직접 접근이 가능하다.
-      await deletePost(this.postItem._id);
-      console.log('deleted');
+      if (confirm('You want to delete it?')) {
+        await deletePost(this.postItem._id);
+        this.$emit('refresh');
+      }
+      // console.log('deleted');
+    },
+    routeEditPage() {
+      const id = `${this.postItem._id}`;
+      this.$router.push(`/post/${id}`);
     },
   },
 };
